Changes to formatting
Changes made to use 100vh and 100vw for the pages.  Made changes to the various files to accommodate for this change.
Changes the delimiter of the theme text files to pipe since some hints included commas.
Changed button animation to move from bottom so it doesn't move over the typing font.